move values to our mission
